# backend/docs/swagger.yaml
basePath: /
definitions:
  domain.AuthStatusResponse:
    properties:
      is_verified:
        type: boolean
    type: object
  domain.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  domain.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  domain.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  domain.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        minLength: 6
        type: string
      otp:
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
  domain.Response:
    properties:
      access_token:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      response_time:
        type: string
      success:
        type: boolean
    type: object
...
  domain.SignupRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
  domain.VerifyEmailRequest:
    properties:
      otp:
        type: string
    required:
    - otp
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the SEASYN backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SEASYN API
  version: "1.0"
paths:
...
  /v1/auth/{provider}/callback:
    get:
      description: Handle OAuth provider callback, verify CSRF state, and return tokens
        as cookies.
      parameters:
      - description: OAuth Provider
        enum:
        - google
        - github
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth Authorization Code
        in: query
        name: code
        required: true
        type: string
      - description: CSRF State
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Invalid state or missing code
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: OAuth exchange failed
          schema:
            $ref: '#/definitions/domain.Response'
      summary: OAuth Callback
      tags:
      - auth
  /v1/auth/{provider}/login:
    get:
      description: |-
        Returns the authorization URL for a specific provider (google, github).
        The frontend should redirect the user to the returned auth_url.
      parameters:
      - description: OAuth Provider
        enum:
        - google
        - github
        in: path
        name: provider
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Unsupported provider
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get OAuth Login URL
      tags:
      - auth
  /v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Update password for an authenticated user
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Change Password
      tags:
      - auth
  /v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send OTP to user email for password reset
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Request Password Reset
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and set access and refresh tokens as HttpOnly
        cookies
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/domain.Response'
      summary: User Login
      tags:
      - auth
  /v1/auth/logout:
    post:
      description: Clear access_token and refresh_token HttpOnly cookies
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Logout User
      tags:
      - auth
  /v1/auth/me:
    get:
      consumes:
      - application/json
      description: Returns 200 OK and verification status if the user has a valid
        access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.AuthStatusResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Check authentication status
      tags:
      - auth
  /v1/auth/otp/send:
    post:
      consumes:
      - application/json
      description: Generates and sends an OTP to the authenticated user's email
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Send OTP
      tags:
      - auth
  /v1/auth/otp/verify:
    post:
      consumes:
      - application/json
      description: Verify the authenticated user's email using an OTP
      parameters:
      - description: Verify Email Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - Bearer: []
      summary: Verify Email with OTP
      tags:
      - auth
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: |-
        Exchange the refresh_token cookie (or request body) for a new access token.
        The server sets new access_token and refresh_token cookies automatically.
      parameters:
      - description: Refresh Request (optional if refresh_token cookie is set)
        in: body
        name: request
        schema:
          $ref: '#/definitions/domain.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Refresh Access Token
      tags:
      - auth
  /v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using an OTP and email
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Reset Password
      tags:
      - auth
  /v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user and receive access and refresh tokens via HttpOnly
        cookies
      parameters:
      - description: Signup Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Email already taken or invalid input
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.Response'
      summary: User Signup
      tags:
      - auth
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"